fear/Archive 사용자 로그인, 회원가입

+++ src/main/java/kr/co/jhta/app/delideli/user/account/domain/UserAddress.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class UserAddress {
-    private Long userAddressKey;
-    private Long userKey;
+    private int userAddressKey;
+    private int userKey;
     private String address;
     private String zipcode;
     private String addrDetail;
